Conflicts:
	CMakeLists.txt
	resources/3rdparty/log4cplus-1.1.2-rc2/CMakeLists.txt

Former-commit-id: 890b34ca8f212478e67549acae3e9c83f85a933f

+++ src/ir/expressions/IntegerConstantExpression.cpp
@@ -24,7 +24,7 @@ namespace storm {
             }
             
             double IntegerConstantExpression::getValueAsDouble(std::pair<std::vector<bool>, std::vector<int_fast64_t>> const* variableValues) const {
-                return getValueAsInt(variableValues);
+                return static_cast<double>(getValueAsInt(variableValues));
             }
             
             int_fast64_t IntegerConstantExpression::getValueAsInt(std::pair<std::vector<bool>, std::vector<int_fast64_t>> const* variableValues) const {
